Renamed settings_db and background_db
