+++ tests/settings.test.js
@@ -1,5 +1,5 @@
 'use strict'
-const GameSettings = require('../src/services/settings_db.cjs')
+const GameSettings = require('../src/services/settingsDb.cjs')
 const { closeAll } = require('../src/services/poolManagement.cjs')
 
 afterAll(() => {
